added check for actual use of lock

+++ Scripts/SequenceProcessor.py
@@ -40,6 +40,7 @@ def run(ref, n, max_step, in_path, out_path):
         out_path = out_path / prefix
         log_path = out_path / 'log.txt'
         lock_path = out_path / '{0}.lock'.format(os.getpid())
+	is_locked = False
 
         #setup and check previous completion
         if not os.path.isdir(out_path):
@@ -51,6 +52,7 @@ def run(ref, n, max_step, in_path, out_path):
                 return
         else:
             lock_path.touch()
+	    is_locked = True
 
         if len(list(out_path.glob('*.lock'))) != 1:
             lock_path.unlink() #some race condition caused there to be two lock files. give up. 
@@ -98,7 +100,8 @@ def run(ref, n, max_step, in_path, out_path):
     except:
         pass
     finally:
-        lock_path.unlink()#remove lock upon completion
+	if is_locked:
+            lock_path.unlink()#remove lock upon completion
 
 
 def runRef(ref_path, out_path):
